Make carret movable and dynamic finally
The solution is a bit tricky. In the core, it just tracks carret position with Selection API.
But the main difficulty is to make carret size equal to a single character. Solution is to add
single character to the div and make it transparent. Since monospace is used, it makes sure
that size of the carret matches font size dynamically even when font size changes from
browser to browser.