Fix carret shift when selection is switched to the span element
Usually, when editing happens, selection API focuses on the text node. However, when ctrl+a is used,
it selects the whole span. Relying strictly on selected component seems too upredictable, so recalculation
is done on input change now. Thus, even if different component selected, but input change, we make sure
that carret is properly shifted.